+++ lib/index.d.ts
@@ -5,6 +5,7 @@ import { ImageMakerService } from './services/image-maker';
 import { PhotoTransformerService } from './services/photo-transformer';
 import { UpscaleService } from './services/upscale';
 import { BackgroundRemoverService } from './services/background-removal';
+import { StyleService } from './services/style';
 export declare class Ablo {
     private axios;
     readonly storage: StorageService;
@@ -13,5 +14,6 @@ export declare class Ablo {
     readonly fontMaker: FontMakerService;
     readonly upscale: UpscaleService;
     readonly removeBackground: BackgroundRemoverService;
+    readonly style: StyleService;
     constructor(apiKey: string, options?: IAbloOptions);
 }
