+++ lib/index.js
@@ -8,6 +8,7 @@ const image_maker_1 = require("./services/image-maker");
 const photo_transformer_1 = require("./services/photo-transformer");
 const upscale_1 = require("./services/upscale");
 const background_removal_1 = require("./services/background-removal");
+const style_1 = require("./services/style");
 class Ablo {
     constructor(apiKey, options = {}) {
         this.axios = axios_1.default.create({
@@ -23,6 +24,7 @@ class Ablo {
         this.fontMaker = new fontmaker_1.FontMakerService(this.axios);
         this.upscale = new upscale_1.UpscaleService(this.axios);
         this.removeBackground = new background_removal_1.BackgroundRemoverService(this.axios);
+        this.style = new style_1.StyleService(this.axios);
     }
 }
 exports.Ablo = Ablo;
