+++ lib/services/style/index.js
@@ -18,7 +18,7 @@ class StyleService {
          * @param params - The parameters for creating a custom style.
          * @returns A Promise that resolves to the created style.
          */
-        this.createCustomStyle = (params) => __awaiter(this, void 0, void 0, function* () {
+        this.createCustom = (params) => __awaiter(this, void 0, void 0, function* () {
             const { data } = yield this.axios.post(`/${StyleService.ENTITY}`, params, {
                 headers: {
                     'Content-Type': 'multipart/form-data',
@@ -30,7 +30,7 @@ class StyleService {
          * Retrieves all custom styles.
          * @returns A Promise that resolves to an array of styles.
          */
-        this.getCustomStyles = () => __awaiter(this, void 0, void 0, function* () {
+        this.getCustom = () => __awaiter(this, void 0, void 0, function* () {
             const { data } = yield this.axios.get(`/${StyleService.ENTITY}/${StyleService.CUSTOM}`);
             return data;
         });
