+++ src/services/style/index.ts
@@ -1,27 +1,11 @@
 import { AxiosInstance } from 'axios'
 import { IStyle } from './style.interface'
-import { ICreateCustomStyleRequest } from './create-custom-style-request.interface'
 
 export class StyleService {
   private static ENTITY = 'styles'
   private static CUSTOM = 'custom'
 
   constructor(private readonly axios: AxiosInstance) {}
-
-  /**
-   * Kicks off the process of creating a custom style.
-   * @param params - The parameters for creating a custom style.
-   * @returns A Promise that resolves to the created style.
-   */
-  createCustom = async (params: ICreateCustomStyleRequest) => {
-    const { data } = await this.axios.post(`/${StyleService.ENTITY}`, params, {
-      headers: {
-        'Content-Type': 'multipart/form-data',
-      },
-    })
-
-    return data
-  }
 
   /**
    * Retrieves all custom styles.
